MSTM

MSTM.SpecialFunctions.cricbessel — Method

cricbessel(n, ds)

Compute the Riccati-Bessel function, return complex values.

MSTM.SpecialFunctions.cspherebessel — Method

cspherebessel(n, z)

Compute spherical Bessel functions jn(z) & yn(z)

MSTM.SpecialFunctions.eulerrotation — Method

eulerrotation(xp)

Euler rotation of a point (xp[1], xp[2], xp[3]).

MSTM.SpecialFunctions.gaussianbeamcoef — Method

gaussianbeamcoef(ctx, α, β, cbeam, nodr)

Regular VSWF expansion coefficients for a gaussian beam, localized approximation.

cbeam = 1/(k omega)

MSTM.SpecialFunctions.gentrancoef — Method

gentrancoef(ctx, itype, xptran, ri, nrow0, nrow1, ncol0, ncol1, iaddrow0, iaddcol0)

Calculates the vwh translation coefficients for a general translation from one origin to another.

MSTM.SpecialFunctions.msta1 — Method

msta1(x, mp)

Determine the starting point for backward recurrence such that the magnitude of Jn(x) at that point is about 10^(-MP).

MSTM.SpecialFunctions.pifunc — Method

pifunc(cb, ephi, nmax, ndim)

Vector spherical harmonic function.

TODO: ndim seems to be useless. Can it be safely deleted?

MSTM.SpecialFunctions.ricbessel — Method

ricbessel(n, ds, ϵ, Ψ)

Compute the Ricatti-Bessel function.

References:

Riccati-Bessel Functions at Wolfram MathWorld

MSTM.SpecialFunctions.sphereplanewavecoef — Method

sphereplanewavecoef(ctx, nsphere, neqns, nodr, nodrmax, α, β, rpos, hostsphere, numberfieldexp, rimedium)

Plane wave expansion coefficients at sphere origins. A phase shift is used.

MSTM.SpecialFunctions.taufunc — Method

taufunc(ctx, cb, nmax)

Normalized vector spherical harmonic function.

MSTM.SpecialFunctions.tranordertest — Method

tranordertest(ctx, r, ri, lmax, ϵ)

Test to determine convergence of regular VSWF addition theorem for max order lmax and translation distance r w/ refractive index ri.

MSTM.SpecialFunctions.vcfuncuprec! — Method

vcfuncuprec(ctx, m, n, k, l, wmax, vcn)

Upwards VC coefficient recurrence, used internally by vcfunc(ctx, m, n, k, l).

MSTM.Data.MSTMConfig — Method

MSTMConfig(filename::String)

Read configurations from a .toml file.

MSTM.Data.TMaxtrix — Method

TMaxtrix(filename::String)

Read a previously calculated T matrix.

MSTM.Mie.mieoa — Method

mieoa(x, ri, nodr0, qeps, ri_medium)

Optically active Lorenz/Mie coefficients

Index

MSTM.Data.MSTMConfig

MSTM.Data.TMaxtrix

MSTM.Mie.mieoa

MSTM.SpecialFunctions.cricbessel

MSTM.SpecialFunctions.cspherebessel

MSTM.SpecialFunctions.eulerrotation

MSTM.SpecialFunctions.gaussianbeamcoef

MSTM.SpecialFunctions.gentrancoef

MSTM.SpecialFunctions.msta1

MSTM.SpecialFunctions.pifunc

MSTM.SpecialFunctions.ricbessel

MSTM.SpecialFunctions.sphereplanewavecoef

MSTM.SpecialFunctions.taufunc

MSTM.SpecialFunctions.tranordertest

MSTM.SpecialFunctions.vcfuncuprec!